+++ src/isEventized.ts
@@ -1,7 +1,7 @@
 import {NAMESPACE} from './constants';
 import type {EventizeGuard, EventizedObject} from './types';
 
-export const isEventized: EventizeGuard = <T extends object>(
-  obj: T,
-): obj is T & EventizedObject =>
+export const isEventized: EventizeGuard = (
+  obj: unknown,
+): obj is EventizedObject =>
   Boolean(obj && (obj as Record<symbol, unknown>)[NAMESPACE]);